+++ CHANGELOG.md
@@ -2,16 +2,19 @@
 
 This file follows the sintaxis defined by [Semantic Versioning](http://semver.org/).
 
-<!-- ## WIP -->
+## WIP
 ## 1.1.00 - 2023-06-10
 
 ### Added
 
 - Utils package added
+- init_from method on DictionaryManager to initiate keywords from a valid one
 
 ### Fixed
 
 - Allow popups by default on BrowserManager
+- Extending keyword method on DictionaryManager now take the current keyword from the DictionaryManager instead from arguments
+- Bug in resolveCaptcha method of BrowserManager that makes browser detect captcha present when isn't one
 
 ## 1.0.00 - 2023-06-01